+++ src/contas/controller/contas.controller.ts
@@ -1,53 +1,47 @@
 import {
   Controller,
-  Get,
   Post,
   Body,
-  Patch,
   Param,
   Delete,
   ParseIntPipe,
+  Get,
 } from '@nestjs/common';
 import { ContasService } from '../services/contas.service';
-import { CreateContaDto } from '../dto/create/create-conta.dto';
-import { UpdateContaDto } from '../dto/update/update-conta.dto';
-import {
-  Funcionario,
-  TipoCargo,
-} from 'src/funcionarios/entities/funcionario.entity';
+import { Funcionario } from 'src/funcionarios/entities/funcionario.entity';
+import { CreateContaCorrenteDto } from '../dto/create/create-conta-corrente.dto';
+import { CreateContaPoupancaDto } from '../dto/create/create-conta-poupana.dto';
+import { TipoCargo } from 'src/common/enums/tipo-.conta.enum';
 
 @Controller('contas')
 export class ContasController {
   constructor(private readonly contasService: ContasService) {}
 
   @Post()
-  criarConta(@Body() createContaDto: CreateContaDto) {
+  criarContaCorrente(@Body() createContaDto: CreateContaCorrenteDto) {
     const funcionario: Funcionario = {
       id: 6,
       nomeFuncionario: 'Nome',
       cargo: TipoCargo.GERENTE,
       telefones: ['123456789'],
     };
-    return this.contasService.abrirConta(funcionario, createContaDto);
+    return this.contasService.abrirContaCorrente(funcionario, createContaDto);
+  }
+
+  @Post()
+  criarContaPoupanca(@Body() createContaDto: CreateContaPoupancaDto) {
+    const funcionario: Funcionario = {
+      id: 6,
+      nomeFuncionario: 'Nome',
+      cargo: TipoCargo.GERENTE,
+      telefones: ['123456789'],
+    };
+    return this.contasService.abrirContaPoupanca(funcionario, createContaDto);
   }
 
   @Get(':id')
   findById(@Param('id') id: string) {
     return this.contasService.findById(+id);
-  }
-
-  @Patch(':id')
-  atualizarConta(
-    @Param('id') id: string,
-    @Body() updateContaDto: UpdateContaDto,
-  ) {
-    const funcionario: Funcionario = {
-      id: 6,
-      nomeFuncionario: 'Nome',
-      cargo: TipoCargo.GERENTE,
-      telefones: ['123456789'],
-    };
-    return this.contasService.atualizarConta(+id, funcionario, updateContaDto);
   }
 
   @Delete(':id')
@@ -61,11 +55,6 @@ export class ContasController {
     return this.contasService.fecharConta(+id, funcionario);
   }
 
-  @Post('depositar/:id')
-  depositar(@Param('id') id: string, @Body('valor') valor: number) {
-    return this.contasService.depositar(+id, valor);
-  }
-
   @Post('sacar/:id')
   sacar(@Param('id') id: string, @Body('valor') valor: number) {
     return this.contasService.sacar(+id, valor);
@@ -74,19 +63,41 @@ export class ContasController {
   @Post('transferir')
   transferir(
     @Body('idOrigem', ParseIntPipe) idOrigem: number,
-    @Body('idDestino',ParseIntPipe) idDestino: number,
+    @Body('idDestino', ParseIntPipe) idDestino: number,
     @Body('valor') valor: number,
   ) {
     return this.contasService.transferir(idOrigem, idDestino, valor);
   }
+}
 
-  @Post('pagar/:id')
+
+ /*  @Post('depositar/:id')
+  depositar(@Param('id') id: string, @Body('valor') valor: number) {
+    return this.contasService.depositar(+id, valor);
+  } */
+
+  /* @Post('pagar/:id')
   pagarConta(@Param('id') id: string, @Body('valor') valor: number) {
     return this.contasService.pagarConta(+id, valor);
-  }
+  } */
 
-  @Get('saldo/:id')
+  /* @Get('saldo/:id')
   consultarSaldo(@Param('id') id: string) {
     return this.contasService.consultarSaldo(+id);
+  } */
+
+
+  /* @Patch(':id')
+  atualizarConta(
+    @Param('id') id: string,
+    @Body() updateContaDto: UpdateContaDto,
+  ) {
+    const funcionario: Funcionario = {
+      id: 6,
+      nomeFuncionario: 'Nome',
+      cargo: TipoCargo.GERENTE,
+      telefones: ['123456789'],
+    };
+    return this.contasService.atualizarConta(+id, funcionario, updateContaDto);
   }
-}
+ */
